backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
supabase>=2.10.0
redis[hiredis]>=5.2.0
httpx>=0.27.0
websockets>=14.0
pyjwt[crypto]>=2.9.0
cryptography>=43.0.0
google-genai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
